Move Alignment definition into its own file
